requirements.txt
feedparser
notion-client
openai
python-dateutil
python-dotenv
httpx
flask
flask-cors
